Fix and test bug in files export
